pyproject.toml
[tool.poetry]
name = "get-poster"
version = "0.1.0"
description = ""
authors = ["Jean-Léon <92361407+JeanLeonHenry@users.noreply.github.com>"]
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pydantic = "^2.7.4"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
